save_json: disallow NaN
Do not allow NaN values in exported JSON to be compliant with JSON
specifications. Prevents users from running into a SyntaxError when
trying to load the JSON in a JavaScript app.

This will raise a ValueError if the provided output does contain nan,
inf, or -inf values. These values should be handled upstream or within
the `EvofrEncoder`.