Fix EvofrEncoder: remove `np.isnan` check
This lead to a ValueError when the passed obj is an np.ndarray type.

I'm completely removing this check because this implementation would
not work as expected in the first place. Since the JSON encoder knows
how to serialize np.nan, the custom encoder's `default` method doesn't
even get called! See detailed comments in #6.