`pivot` and `var_names` incompatible in `VariantFrequencies`
Previously you could specify both `pivot` and `var_names` when
instantiating `VariantFrequencies`. But if this was done, the value
of `pivot` was silently ignored in favor of making the pivot variant
the last name in `var_names`.

This pull request fixes this misleading behavior by making those
parameters mutually exclusive, and raising an error if the user
tries to set both.