Infer date interval for forecast dates
Instead of assuming that forecasts should be produced at daily
intervals, inspect the interval between the given input dates and use
that interval to generate forecasts. This commit adds unit tests for
expected behavior and then implements that behavior.

Note this fix only works when the input dates use a fixed size interval
in units of days. If the interval between dates is variable (e.g.,
monthly), then forecasts will be based on the interval between the last
two dates in the given input.